Rework Batch Files: Make them run from double-click, safe search for DXLSTDLIB directory, rename setupDOORS
The existing command for determining the DXL Std Lib Dir was working with %CD% which does not work, when started from eclipse.
The new way of determining the path by going from the path of the current file upwards is much better.

setupDOORS had to be renamed to configureDOORS because Windows 7 UAC seems to grant admin to ALL files that begin with setup....exe (lol)

make_release.bat has to be still reworked!! make_clean has changed its functionality.